pyproject.toml
[project]
name = "odo-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.50.0",
    "pushbullet-py>=0.12.0",
    "python-dotenv>=1.0.1",
]
